+++ examples/generate_key/main.go
@@ -6,9 +6,10 @@ import (
 )
 
 func main() {
-	secret := hedera.GenerateSecretKey()
-	fmt.Printf("secret = %v\n", secret)
+	secret, mnemonic := hedera.GenerateSecretKey()
+	fmt.Printf("secret   = %v\n", secret)
+	fmt.Printf("mnemonic = %v\n", mnemonic)
 
 	public := secret.Public()
-	fmt.Printf("public = %v\n", public)
+	fmt.Printf("public   = %v\n", public)
 }