+++ CHANGELOG.md
@@ -4,7 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.2.0] - 2018-12-07
+## Unreleased
+
+### Changed
+
+ - `hedera.GenerateSecretKey` now returns both the secret key and a 24-word mnemonic that can be used to recover the secret key.
+
+### Added
+
+ - `hedera.GenerateSecretKeyWithPassword` functions exactly as `hedera.GenerateSecretKey` except the
+   returned mnemonic now will require the password when used to recover the secret key
+
+## 0.2.0 - 2018-12-07
 
 ### Fixed
 
@@ -33,6 +44,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - For transactions, `.Execute()` now returns the ID directly instead of being wrapped
    in a 1-element struct.
 
-## [0.1.0] - 2018-11-29
+## 0.1.0 - 2018-11-29
 
 Initial stable release.